WIP sw handling
