+++ src/components/common/Notification/index.js
@@ -1,50 +1,17 @@
 import React, { useState, useEffect } from 'react';
+import PropTypes from 'prop-types';
 import D from 'i18n';
-import * as serviceWorker from 'serviceWorker';
 import './notification.scss';
 
-const Notification = ({ setQueenSwState, queenSwState }) => {
-  const [init, setInit] = useState(queenSwState);
-  const [open, setOpen] = useState(false);
-  const [installingServiceWorker, setInstallingServiceWorker] = useState(false);
-  const [waitingServiceWorker, setWaitingServiceWorker] = useState(null);
-  const [isUpdateAvailable, setUpdateAvailable] = useState(false);
-  const [isServiceWorkerInstalled, setServiceWorkerInstalled] = useState(false);
+const Notification = ({ serviceWorkerInfo }) => {
+  const [open, setOpen] = useState(true);
 
-  useEffect(() => {
-    console.log('init value : ', init);
-    if (!init) {
-      const install = async () => {
-        const configuration = await fetch(`${window.location.origin}/configuration.json`);
-        const { QUEEN_URL } = await configuration.json();
-        serviceWorker.register({
-          QUEEN_URL,
-          onInstalling: installing => {
-            setInstallingServiceWorker(installing);
-            setOpen(true);
-          },
-          onUpdate: registration => {
-            setWaitingServiceWorker(registration.waiting);
-            setUpdateAvailable(true);
-            setOpen(true);
-          },
-          onWaiting: waiting => {
-            setWaitingServiceWorker(waiting);
-            setUpdateAvailable(true);
-            setOpen(true);
-          },
-          onSuccess: registration => {
-            setInstallingServiceWorker(false);
-            setServiceWorkerInstalled(!!registration);
-            setOpen(true);
-          },
-        });
-        setQueenSwState(true);
-        setInit(true);
-      };
-      install();
-    }
-  }, [init, setQueenSwState]);
+  const {
+    installingServiceWorker,
+    waitingServiceWorker,
+    isUpdateAvailable,
+    isServiceWorkerInstalled,
+  } = serviceWorkerInfo;
 
   const updateAssets = () => {
     if (waitingServiceWorker) {
@@ -97,3 +64,11 @@ const Notification = ({ setQueenSwState, queenSwState }) => {
 };
 
 export default Notification;
+/* Notification.propTypes = {
+  serviceWorkerInfo: PropTypes.shape({
+    installingServiceWorker: PropTypes.shape({}).isRequired,
+    waitingServiceWorker: PropTypes.shape({}).isRequired,
+    isUpdateAvailable: PropTypes.shape({}).isRequired,
+    isServiceWorkerInstalled: PropTypes.shape({}).isRequired,
+  }).isRequired,
+}; */
